week_7_project/2023/requirements.txt
kafka-python==1.4.6
confluent_kafka
requests
avro
faust
fastavro
json